Removes deprecated ruin blacklists and other .txt files. (#2441)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Removes the deprecated ruin blacklists and some text files that are no
longer in use. Changes the comment in the map datum files to point
people to the map catalogue instead so they know to add their ruin to it
now.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I was asked by latency to remove the deprecated ruin blacklists so
people would stop editing them, and mark pointed me to a few other
deprecated files in the same area that can go with them.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: Ruin blacklist text files.
del: Misc. deprecated text files.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->