Quick tweaks to syndicate clothes and the Hunter Montagne (#2460)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Title


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/84818adc-97a7-4929-94bf-92a1e2a45458)

![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/6c132c53-14da-4af7-8bb8-68ba77af1169)

![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/3ae13eb3-0d82-41e2-a54f-f5bd64cd68be)

This PR also does some minor tweaks to the nanotrasen coat, hehe.

!!SURPRISE BONUS!!

This PR also resprites the syndicate mask, killing one of the oldest
sprites in the game. Credits to @axietheaxolotl for the original sprite.

![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/2348ef11-272a-43e4-864b-c74ddfd6c2d9)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Mark my words, the Syndicate will have cohesion eventually. But for now,
cruft removal.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy, AxieTheAxolotl
imageadd: The syndicate operator vest and helmet, the syndicate
captain's garb, and the hunter montagne's garb have been adjusted with
better colors.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->