Fixes bloodstains getting stuck (#2456)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![image](https://github.com/shiptest-ss13/Shiptest/assets/45698967/0cfb95d1-673a-47a9-94dc-af3fba94430a)
As shown in image, bloodstains were getting stuck in the shape of the
first item to draw blood that round

Ports the rest of
[https://github.com/BeeStation/BeeStation-Hornet/pull/3111](url), which
seems to fix it.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Blub :)

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: bloodstains no longer steal the shape of the first item to spill
blood
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->