The Cybersun-ening (#2484)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR changes the outfits of several jobs related to cybersun, as well
as vastly overhauling their aesthetics.


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/cf3c8167-77ae-44d4-b7d0-f652efabdd1a)

This PR also adds completely new items, such as:

- The Cybersun Officer's Cap
- The Cybersun Officer's Suit
- The Cybersun Captain's Haori
- The Cybersun Agent's Jumpsuit
- The Cybersun Agent's Cap
- The Cybersun Trauma Team Vest
- The Cybersun Trauma Team Helmet
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Helps balkanize the syndicate further aesthetically, which is the
current direction we're going for the faction as a whole.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
:cl: PositiveEntropy
fix: Cybersun outfits have been adjusted to be less syndicate-generalist
and more Cybersun.
imageadd: Cybersun has received a full visual overhaul, as well as new
goodies!

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->