hades missing sprites for right hand in-hands fix (#2559)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

The hades rifle didn't have a right-hand sprite for its mostly-fully
charged inhands
There doesn't seem to be any difference between the sprites for charge
levels 4 and 5 for left hand inhands, so I just copied the sprites for
charge level 4 for this


![image](https://github.com/shiptest-ss13/Shiptest/assets/24857008/06072070-5275-4efb-a9e7-f715598c82dd)

## Why It's Good For The Game

fixes #2527

## Changelog

🆑 
fix: The hades energy gun no longer becomes invisible sometimes
/🆑

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->