Adjusts BYOND Skin Buttons (#2570)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

### BEFORE


![image](https://github.com/shiptest-ss13/Shiptest/assets/29362068/734632d0-8008-42cd-a545-7c1ed1010948)

### AFTER


![image](https://github.com/shiptest-ss13/Shiptest/assets/29362068/ea499d7a-e0d4-4d9d-8f1a-d3c5fb29981f)

Makes the bar at the top-right of the screen much more consistent, as
well as swaps the "Forum" button for a "Lore" button, as we don't have a
forum and our lore primer is important. Also shrinks the words used on
the buttons so it looks a little bit better on widescreen mode.

Don't ask me to find out what's going on with the report issues button.

## Why It's Good For The Game

Makes buttons that are always visible a little bit less infuriatingly
unbalanced, and puts a focus on the lore primer.

## Changelog

:cl:
tweak: Adjusts spacing of the top-right BYOND window buttons
add: Adds a lore button to the top-right
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->