Adds what seems to be a missing pixel to the skillcapes and fixes a dead pixel on the cleaning skillcape (#2593)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Each skillcape was missing a pixel on the left side of the front view,
which becomes obvious with the cleaning/mining ones which have a pulsing
animation

Also removes a white spot on the cleaning cape

## Why It's Good For The Game

fixes #1168

## Changelog

:cl:
fix: obliterated a floating white pixel on the cleaning skillcape
fix: skillcapes are no longer missing a pixel on their left side of the
front
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->