Nerfs legion core implanting to not be an aheal (#2590)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
The legion core implant now gives you a potent heal of all four damage
types (-100 brute -100 burn -50 tox -50 oxy) instead of a literal aheal.
It now also deals 10 clone damage as a drawback to organics/FBPS (IPCS
excluded because they cant use clone damage medicine).
Due to how adjustBruteLoss and adjustBurnLoss work, the implanted core
no longer heals mechanical bodyparts, making it mostly useless for IPCs
and FBPs only healing oxygen and toxin damage.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
1. Player accessible aheals are not good and encourage exploiting to
cure ailments or gain an advantage.
2. Synthetics could use self-surgery to implant legion cores on the go
for a safety net heal. While not necessarily bad, it was insanely
powerful as an aheal and negated the requirement of stabilizing the core
and getting another person to put it in you.
3. It had literally no drawbacks. A strong consumable healing ability is
cool, but it should come with a cost.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: legion core implanting no longer aheals you on use
add: legion core implant now just does a potent organic heal with minor
clone damage when used
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: zevo <95449138+Zevotech@users.noreply.github.com>
Co-authored-by: Mark Suckerberg <mark@suckerberg.gay>