Throws a fax machine on the Paracelsus class. (#2630)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![image](https://github.com/shiptest-ss13/Shiptest/assets/95449138/14347020-eb16-4623-9d1a-055d890c39d2)
does this.
also rotates the bedsheet in the captains bedroom to be the right
direction
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It ain't solgov gameplay without a fax machine. Also provides another
way to communicate with the vessel for medical emergencies
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: adds a fax machine to the paracelsus-class
fix: rotates a bedsheet on the paracelsus class to be the right
direction
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->